requirements.txt
requests
Pillow
numpy
happybase